tele/requirements.txt
python-dotenv
fastapi
uvicorn
pydantic
pydantic-settings
email-validator
python-multipart
httpx
pytest
pytest-asyncio
redis[hiredis]
dadata
dnspython
